fix(app): match new session panel corners
