+++ packages/app/e2e/tsconfig.json
@@ -10,7 +10,7 @@
     "./performance/timeline-stability/fixture.test.ts",
     "./performance/timeline-stability/fixture.ts",
     "./performance/unit/visual-stability.test.ts",
-    "./regression/new-session-window-corner.spec.ts",
+    "./regression/new-session-panel-corner.spec.ts",
     "./regression/session-timeline-context-resize.spec.ts",
     "./utils/**/*.ts"
   ]
